requirements.txt
Django>=5.1,<5.2
Pillow>=10.0.0
pdf2image>=1.16.0
python-decouple>=3.8
